Move Deposits.upsert into its own service
We need this to avoid a circular dependency where events needs to
reference the Deposits service.
